restructuring the project code
