+++ backend/devsync/projects/views/__init__.py
@@ -1,0 +1,4 @@
+from .project import ProjectViewSet
+from .member import ProjectMemberViewSet
+from .department import DepartmentViewSet
+from .invitation import ProjectInvitationViewSet
