+++ backend/devsync/roles/renderers.py
@@ -1,0 +1,5 @@
+from config.utils.renderers import ListRenderer
+
+
+class RoleListRenderer(ListRenderer):
+    wrapper_key = 'roles'
